+++ src/output/output_srt.cpp
@@ -28,7 +28,7 @@ namespace Mist {
   
   void OutProgressiveSRT::sendNext(){
     char * dataPointer = 0;
-    unsigned int len = 0;
+    size_t len = 0;
     thisPacket.getString("data", dataPointer, len);
     //ignore empty subs
     if (len == 0 || (len == 1 && dataPointer[0] == ' ')){
